pyproject.toml
[project]
name = "dux"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.3",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "passlib[argon2]>=1.7.0",
    "alembic>=1.13.0",
    "itsdangerous>=2.1.0",
    "webauthn>=2.7.0",
    "slowapi>=0.1.9",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.28.0",
    "pytest>=8.0.0",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "python-dotenv>=1.2.1",
    "pypdf>=4.0.0",
    "python-docx>=0.8.11",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "pypdf2>=3.0.1",
    "openai>=1.0.0",
    "pdf2image>=1.17.0",
    "langchain-huggingface>=1.2.0",
    "sentence-transformers>=5.2.0",
]

[tool.setuptools]
packages = ["server"]
